MDEV-14822 binlog.binlog_killed fails with wrong result
Problem was timing between the thread that was killed and reading the
binary log.

Updated the test to wait until the killed thread was properly terminated
before checking what's in the binary log.

To make check safe, I changed "threads_connected" to be updated after
thd::cleanup() is done, to ensure that all binary logs updates are done
before the variable is changed.  This was mainly done to get the
test deterministic and have now other real influence in how the server
works.