Deb: Re-rename *.files into *.install
The correct format for new Debian packages is to define install lists
as *.install files. This has already been fixed for 10.3, but it when
merging commit c5a525bda2b64dd008598afa51f3b7c6347e24fc from 10.2, this
change was overlooked and a file of the old format slipped in.