Deb: Add missing files into correct packages and rest in not-installed
There was a lot of files generated by the sources that were not installed
in any package. This fixes most of those issues, but not all. Files still
outside any package are in the not-installed, which is used by dh_install.

Also make sure all configuration files are installed to the correct
location that matches layout used by Debian official packages.