MDEV-15143 InnoDB: Rollback of trx with id 0 completed
When InnoDB has completed the rollback of a recovered transaction,
it used to display the transaction identifier.

This was broken in MySQL 5.7.2 in
https://github.com/mysql/mysql-server/commit/2f5f3cd3ac46d46c7635e778b338f63b5521cd46
which was merged to MariaDB 10.2.2 in
commit 2e814d4702d71a04388386a9f591d14a35980bfe.

trx_rollback_active(): Cache the transaction ID before it will be
reset by transaction commit. Do not display the message if the
rollback was interrupted by shutdown (MDEV-13797, MDEV-12352).