Fix sporadic failure of MTR test galera.galera_many_tables_pk
The test relies on the false assumption that node_1 is always
"assigned" autoincrement value 1 for a given key.
However that is no necessarily the case when wsrep autoincrement
control is enabled.

This patch changed the test so that it does not make use of a
autoincrement key, which is not fundamental for the test itself.

codership/mysql-wsrep#322