MDEV-14340: mariadb.pc install should be arch dependent
As such use INSTALL_LIBDIR to set the path.

RPMs already have this ready to use.

Debian used DEB_HOST_MULTIARCH to set the CMAKE_CXX_LIBRARY_ARCHITECTURE
which is incidently used for other apects of cross compulation in CMake (added
2.8.11)
https://www.debian.org/doc/manuals/maint-guide/advanced.en.html#multiarch

Despite being a move in location, this is a stable move as
this file is for the use by pkg-config. Pkg-config will pick the
file up in the new location transparently. It also will
correct cross compulation problems that against mariadb.so
because the location is now correct for the achitecture compiled.

Debian Wheezy has 2.8.9 (https://packages.debian.org/wheezy/cmake)
which isn't too important as far as CMAKE_CXX_LIBRARY_ARCHITECTURE
because we're primary using for our own goals.

Also moves/leaves setting of INSTALL_PLUGINDIR and INSTALL_MYSQLTESTDIR
to cmake/install_layout.cmake as was its intended use.