Fix MTR test galera.galera_gcache_recover_manytrx
The kills the two nodes that compose the cluster and then recovers them.
The first node that is recovered is expected to recover its gcache, so
that when the second node is recovered, it is expected that it rejoins
the cluster via IST.
However, it is possible that if the second node is killed while it is
applying in TOI mode, its local state is marked unsafe. In which case,
SST is the only option to rejoin the cluster.

codership/mysql-wsrep#323