innodb: trx_roll_must_shutdown straight to trx_roll_must_shutdown
Suggested by Marko on github pr #576

In fact, we could modify trx_roll_must_shudown() and its sole
caller so that it will abort any transaction rollback on a normal
(non-slow) shutdown. In that way, shutdown would become a non-issue.

The rollback would then be performed on the subsequent startup, in the background.